requirements.txt
fastapi
uvicorn
pymongo
bson
paho-mqtt
